deleted comment in autograde
had to change autograde so that it took no arguments and created the reports in a loop. it will now assume "CSI" is the name of the directory it is supposed to look through. the previous incarnation, which took the name of the directory as an argument, is preffered by me, but the new requirements made me change it